# docs/calculator.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "calculator.proto",
    "description": "The package name should match the Go package for consistency.",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "CalculatorService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/add": {
      "post": {
        "summary": "Add performs addition and maps to a RESTful POST endpoint.",
        "operationId": "CalculatorService_Add",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoCalculationResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "AddRequest defines the structure for an addition RPC call.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoAddRequest"
            }
          }
        ],
        "tags": [
          "CalculatorService"
        ]
      }
    },
    "/v1/divide": {
      "post": {
        "summary": "Divide performs division and maps to a RESTful POST endpoint.",
        "operationId": "CalculatorService_Divide",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/protoCalculationResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": "DivideRequest defines the structure for a division RPC call.",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/protoDivideRequest"
            }
          }
        ],
        "tags": [
          "CalculatorService"
        ]
      }
    }
  },
  "definitions": {
    "protoAddRequest": {
      "type": "object",
      "properties": {
        "a": {
          "type": "integer",
          "format": "int32"
        },
        "b": {
          "type": "integer",
          "format": "int32"
        }
      },
      "description": "AddRequest defines the structure for an addition RPC call."
    },
    "protoCalculationResponse": {
      "type": "object",
      "properties": {
        "result": {
          "type": "integer",
          "format": "int32"
        }
      },
      "description": "CalculationResponse is the generic response for all calculation RPCs."
    },
    "protoDivideRequest": {
      "type": "object",
      "properties": {
        "dividend": {
          "type": "integer",
          "format": "int32"
        },
        "divisor": {
          "type": "integer",
          "format": "int32"
        }
      },
      "description": "DivideRequest defines the structure for a division RPC call."
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}